requirements.txt
joblib==1.4.2
matplotlib==3.7.1
numpy==1.25.0
pandas==2.0.2
Pillow==11.1.0
scikit_learn==1.6
scipy==1.10.1
torchsummary==1.5.1
tqdm>=4.66.3
openpyxl==3.1.5
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
